+++ src/HSSV/AdjacencyWrapper.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <Defines.h>
+
 #include <glm/glm.hpp>
 #include <glm/gtc/type_ptr.hpp>
 
@@ -7,15 +9,15 @@
 
 class Adjacency;
 
-const float* getVertex3v(Adjacency const* ad, size_t vertexID);
+const float* getVertex3v(Adjacency const* ad, u32 vertexID);
 
-const glm::vec3& getEdgeVertexLow(Adjacency const* ad, size_t edgeId);
+const glm::vec3& getEdgeVertexLow(Adjacency const* ad, u32 edgeId);
 
-const glm::vec3& getEdgeVertexHigh(Adjacency const* ad, size_t edgeId);
+const glm::vec3& getEdgeVertexHigh(Adjacency const* ad, u32 edgeId);
 
-size_t getNofOppositeVertices(Adjacency const* ad, size_t edgeId);
+u32 getNofOppositeVertices(Adjacency const* ad, u32 edgeId);
 
-const glm::vec3& getOppositeVertex(Adjacency const* ad, size_t edgeId, size_t oppositeVertexID);
+const glm::vec3& getOppositeVertex(Adjacency const* ad, u32 edgeId, u32 oppositeVertexID);
 
-std::vector<glm::vec3> getEdgeOppositeVertices(Adjacency const* ad, size_t edgeId);
+std::vector<glm::vec3> getEdgeOppositeVertices(Adjacency const* ad, u32 edgeId);
